requirements.txt
# Requirements for Robot Framework with SikuliX

# Core Robot Framework
robotframework==7.0.1

# SikuliX Library for Robot Framework
robotframework-SikuliLibrary==2.0.5

# Additional useful libraries
robotframework-seleniumlibrary==6.2.0
robotframework-requests==0.9.5
robotframework-databaselibrary==1.4.4
robotframework-sshlibrary==3.8.0

# Image processing (optional, for advanced image manipulation)
Pillow==10.1.0
opencv-python==4.8.1.78

# Reporting and logging
robotframework-pabot==2.17.0  # For parallel test execution
robotframework-metrics==3.3.3

# Test data management
pyyaml==6.0.1
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
